# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VNSF Scholarship Management API
  version: 0.1.0
  license: { name: Proprietary, identifier: LicenseRef-Proprietary }
servers:
  - url: /api/v1
security: [{ sessionCookie: [] }]
paths:
...
  /students:
    get:
      summary: List students in the caller data scope
      operationId: listStudents
      parameters:
        - {
            in: query,
            name: page,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            in: query,
            name: size,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        "200": { description: Scoped student page }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Create a student after duplicate checks
      operationId: createStudent
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentInput" }
      responses:
        "200": { description: Student created }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /students/{id}:
    get:
      summary: Get a field-filtered student
      operationId: getStudent
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        {
          "200": { description: Masked field-policy detail },
          "404": { $ref: "#/components/responses/NotFound" },
        }
    patch:
      summary: Update allowed student fields
      operationId: updateStudent
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/StudentPatch" } },
          }
      responses:
        "200": { description: Updated student }
        "412": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
  /students/duplicates/check:
    get:
      summary: Find scoped exact duplicate candidates by normalized name and birth date
      operationId: checkStudentDuplicates
      parameters:
        - {
            in: query,
            name: full_name,
            required: true,
            schema: { type: string, minLength: 2, maxLength: 150 },
          }
        - {
            in: query,
            name: date_of_birth,
            required: true,
            schema: { type: string, format: date },
          }
      responses:
        "200": { description: Scoped duplicate candidates }
        "404": { $ref: "#/components/responses/NotFound" }
  /students/{id}/guardians:
    parameters: [{ $ref: "#/components/parameters/ResourceId" }]
    get:
      summary: List guardians with masked contact details
      operationId: listStudentGuardians
      responses:
        "200": { description: Guardian list with masked contacts }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Add an encrypted guardian contact
      operationId: addStudentGuardian
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/GuardianInput" } },
          }
      responses:
        "200": { description: Guardian created with masked contact }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /students/{id}/school-history:
    get:
      summary: List immutable school assignments for a scoped student
      operationId: listStudentSchoolHistory
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        "200": { description: Effective-dated school history }
        "404": { $ref: "#/components/responses/NotFound" }
  /students/{id}/school-transfer:
    post:
      summary: Atomically transfer a student and append history/profile versions
      operationId: transferStudentSchool
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  { $ref: "#/components/schemas/StudentSchoolTransferInput" },
              },
          }
      responses:
        "200": { description: Transferred student }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
...
  /manual-transfers:
    get:
      summary: List transfers in the caller's program, school or student scope
      operationId: listManualTransfers
      responses:
        "200": { description: Scoped transfer page }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Record an externally executed transfer
      operationId: createManualTransfer
      description: Records an external transfer; never initiates payment.
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/ManualTransferInput" },
                },
            },
        }
      responses:
        "200": { description: Awaiting student confirmation }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /manual-transfers/{id}:
    get:
      summary: Read a scoped transfer with confirmation and correction evidence
      operationId: getManualTransfer
      parameters: [{ $ref: "#/components/parameters/ResourceId" }]
      responses:
        "200": { description: Transfer detail }
        "404": { $ref: "#/components/responses/NotFound" }
  /manual-transfers/{id}/confirm:
    post:
      summary: Confirm receipt or report non-receipt as the linked student
      operationId: confirmManualTransfer
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferConfirmationInput" }
      responses:
        "200": { description: Confirmation recorded append-only }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
  /manual-transfers/{id}/corrections:
    post:
      summary: Mark the original corrected and create an immutable replacement
      operationId: correctManualTransfer
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/ResourceId" }
        - { $ref: "#/components/parameters/IfMatch" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransferCorrectionInput" }
      responses:
        "200": { description: Replacement transfer created }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/Conflict" }
        "428": { $ref: "#/components/responses/PreconditionRequired" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: vnsf_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  parameters:
...
    ResourceId:
      {
        in: path,
        name: id,
        required: true,
        schema: { type: string, format: uuid },
      }
    IfMatch:
      {
        in: header,
        name: If-Match,
        required: true,
        schema: { type: string },
        example: '"3"',
      }
    IdempotencyKey:
      {
        in: header,
        name: Idempotency-Key,
        required: true,
        schema: { type: string, minLength: 16, maxLength: 128 },
      }
  schemas:
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string }
        params: { type: object, additionalProperties: true }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message_key, field_errors, correlation_id]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message_key: { type: string, example: errors.versionConflict }
        field_errors:
          { type: array, items: { $ref: "#/components/schemas/FieldError" } }
        correlation_id: { type: string, format: uuid }
    StudentInput:
      type: object
      additionalProperties: false
      required:
        [student_code, full_name, date_of_birth, program_id, current_school_id]
      properties:
        student_code: { type: string, minLength: 3, maxLength: 30 }
        full_name: { type: string, minLength: 1, maxLength: 150 }
        date_of_birth: { type: string, format: date }
        program_id: { type: string, format: uuid }
        current_school_id: { type: string, format: uuid }
        grade_level_current: { type: integer, minimum: 1, maximum: 12 }
        duplicate_override_reason:
          { type: string, minLength: 10, maxLength: 500 }
    StudentPatch:
      type: object
      additionalProperties: false
      properties:
        full_name: { type: string, minLength: 1, maxLength: 150 }
        date_of_birth: { type: string, format: date }
        grade_level_current: { type: integer, minimum: 1, maximum: 12 }
    GuardianInput:
      type: object
      additionalProperties: false
      required: [full_name, relationship]
      anyOf: [{ required: [phone] }, { required: [email] }]
      properties:
        full_name: { type: string, minLength: 2, maxLength: 150 }
        relationship: { type: string, enum: [MOTHER, FATHER, GUARDIAN, OTHER] }
        is_primary: { type: boolean, default: false }
        phone: { type: string, minLength: 8, maxLength: 20 }
        email: { type: string, format: email }
    StudentSchoolTransferInput:
      type: object
      additionalProperties: false
      required: [target_school_id, effective_from, reason]
      properties:
        target_school_id: { type: string, format: uuid }
        effective_from: { type: string, format: date }
        reason: { type: string, minLength: 10, maxLength: 500 }
...
    ManualTransferInput:
      type: object
      additionalProperties: false
      required:
        [
          student_id,
          period_id,
          transfer_type,
          amount,
          currency,
          transferred_at,
          reference,
        ]
      properties:
        student_id: { type: string, format: uuid }
        period_id: { type: string, format: uuid }
        transfer_type: { type: string }
        amount: { type: string, pattern: '^\\d+(\\.\\d{1,2})?$' }
        currency: { type: string, enum: [VND, USD] }
        transferred_at: { type: string, format: date-time }
        reference: { type: string }
    TransferConfirmationInput:
      type: object
      additionalProperties: false
      required: [result]
      properties:
        result: { type: string, enum: [RECEIVED, NOT_RECEIVED] }
        note: { type: string, maxLength: 1000 }
    TransferCorrectionInput:
      type: object
      additionalProperties: false
      required:
        [
          transfer_type,
          amount,
          currency,
          transferred_at,
          reference,
          reason_code,
          reason,
        ]
      properties:
        transfer_type: { type: string, minLength: 1, maxLength: 30 }
        amount: { type: string, pattern: '^\d+(\.\d{1,2})?$' }
        currency: { type: string, enum: [VND, USD] }
        transferred_at: { type: string, format: date-time }
        reference: { type: string, minLength: 1, maxLength: 100 }
        reason_code: { type: string, minLength: 2, maxLength: 50 }
        reason: { type: string, minLength: 10, maxLength: 1000 }
...
  responses:
...
    Unauthorized:
      {
        description: Authentication or MFA required,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    NotFound:
      {
        description: Missing or concealed by scope,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    Conflict:
      {
        description: "Duplicate, state or version conflict",
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    ValidationFailed:
      {
        description: Business validation failed,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
    PreconditionRequired:
      {
        description: If-Match/version is required,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorResponse" } },
          },
      }
...